+++ webScraping/webScraping/Yahoo.cs
@@ -7,9 +7,9 @@ namespace webScraping
 {
     class Yahoo : SearchEngine
     {
-        public override async Task<int> GetResult(string input)
+        public override async Task<long> GetResult(string input)
         {
-            int quantity = 0;
+            long quantity = 0;
             string url = "https://search.yahoo.com/search?p=" + input;
             string keyWord = "Next</a><span>";
 
@@ -21,7 +21,7 @@ namespace webScraping
             string result = html.Substring(start, end - start);
 
             string onlyNumber = Regex.Replace(result, "[a-zA-Z</>,\n. ]", "");
-            quantity = Convert.ToInt32(onlyNumber);
+            quantity = Convert.ToInt64(onlyNumber);
             return quantity;
 
         }
